pyproject.toml
[project]
name = "django-smart-env"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-environ>=0.11.2",
]

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.36",
]

[tool.uv]
dev-dependencies = [
    "black>=24.8.0",
    "django>=4.2.16",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "pdbpp>=0.10.3",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3",
]


[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | snapshots
)/
'''
